move helpers to helpers_test
